refactor(controller): 전역 예외 처리를 사용하도록 컨트롤러 리팩토링
