+++ src/main/java/gift/controller/api/ProductController.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package gift.controller.api;
 
 
 import gift.dto.CreateProductRequestDto;
